+++ src/snapshot/resource_checksum_hash.rs
@@ -42,7 +42,7 @@ where
 
         resource.hash(&mut hasher);
 
-        let result = ChecksumPart(hasher.finish());
+        let result = ChecksumPart(hasher.finish() as u128);
 
         if let Ok(mut checksum) = checksum.get_single_mut() {
             *checksum = result;
